Download data for a single export job.

Executes full download process for a single export job.

Usage

download_data(
  qnr_id,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  path,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)

Arguments

qnr_id

Questionnaire ID. Format: qnr_id$version

export_type

Type of data to export. Values: c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")

interview_status

Status of interviews to export. Values: c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")

from

Start date for interviews to export. date-time, UTC

to

End date for interviews to export. date-time, UTC

access_token

Access token to external storage. Relevant only if external cloud storage is destination for the export file.

refresh_token

Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.

storage_type

External storage type, if relevant. Values: c("Dropbox", "OneDrive", "GoogleDrive")

translation_id

Translation ID for variable and value labels to include in export files.

include_meta

Logical. If TRUE, include questionnaire metadata in export file.

path

File path where export file should be downloaded

server

Full server web address (e.g., https://demo.mysurvey.solutions, https://my.domain)

workspace

Character. Name of the workspace whose data to download.

user

API user name

password

API password

Details

Starts export job. Specifies the parameters of the download job through the qnr_id, export_type, interview_status, from, to, access_token, refresh_token, storage_type, translation_id, and include_meta parameters

Gets regular updates on the progress of the export job. Messages about progress.

Downloads file to desired destination. Use the path parameter to specify where the export file should be saved.