Get list of questionnaires with a matching title.

Get a data frame of questionnaires that match the character vector provided.

find_matching_qnrs(
  matches,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)

Arguments

matches
Atomic character vector. Character string to be detected in questionnaire title. May be a string, sub-string, or regex.

workspace
Character. Name of the workspace whose questionnaires to query.

server
Full server web address (e.g., https://demo.mysurvey.solutions, https://my.domain)

user
API user name

password
API password

Value

Data frame of matching questionnaires, if successful; warning message, if not successful.

Details

If matching questionnaires are found, return a data frame with attributes. These attributes can be used, among other things, to find the necessary parameters for downloading any associated data.

If no matching questionnaires are found, return a warning message.

In both cases, print the outcome to the console for information.