Compute the last sync date for each interviewer

Conducts full workflow to get the last sync date for each interviewer:

get_last_sync_dates(
  start,
  end,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)

Arguments

start
Character vector. Date in YYY-MM-DD format

end
Character vector. Date in YYY-MM-DD format

workspace
Character. Workspace whose action logs to get.

server
Full server web address (e.g., https://demo.mysurvey.solutions, https://my.domain)

user
API user name

password
API password

Value

Data frame consisting of one or more user action log.

Details

Obtain the list of all interviewers

Fetch the action logs for each interviewers

Determine the last sync date for each interviewer

Returns a data frame. Columns contain Contains the user ID (UserId) and last sync date (last_sync_date). Rows correspond to interviewers. NA in last_sync_date mean that there was no sync action identified in the action log--that is between the start and end dates used to obtain the action logs.